projects/trello/requirements.txt
Django==5.1.5
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.8
psycopg2==2.9.10
requests==2.32.3
wheel==0.44.0
